newnet: Fix the host mixing up broadcast messages with messsages to the host, by making NetQueue even more message-oriented. Use -Wall -Werror on C++ files too, and fix the corresponding werrors. Don't use gameRand() from scripts, since the scripts are apparently only executed on the host. Add syncDebug() log function, which is only printed, if it would print different things on different clients.

+++ src/wavecast.cpp
@@ -52,7 +52,7 @@ static std::vector<WavecastTile> generateWavecastTable(unsigned radius)
 
 	std::vector<RationalAngle> unsortedAngles;
 
-	for (int diamond = 1; diamond*TILE_UNITS < radius*2; ++diamond)  // Factor is a bit more than needed to surround the circle with diamonds.
+	for (int diamond = 1; (unsigned)diamond*TILE_UNITS < radius*2; ++diamond)  // Factor is a bit more than needed to surround the circle with diamonds.
 	{
 		for (int quadrant = 0; quadrant < 4; ++quadrant)
 		{
